feat(settings): build out Account Settings; split personal vs workspace (CVS-86)
Turn the thin account surface into a proper settings area and resolve the
personal/workspace redundancy the epic calls out.

Account Settings (/settings) now has: Profile, Appearance, Notifications &
preferences, and API keys — the latter three moved here from Workspace
Settings since they're per-user — plus scaffolded Connected accounts and
Billing sections (disabled "Soon" affordances; OAuth pipes are owned by the
Integrations lane, billing pending backend).

Workspace Settings (/settings/workspace) is now workspace-only: data-source
connections + members/roles. Cross-links added between the two pages.

Stacked on CVS-87 (PR #18) — merges after it. Closes CVS-86.
